backend/requirements.txt
# backend/requirements.txt
Django==3.2.12
djangorestframework==3.13.1
djangorestframework-simplejwt
psycopg2-binary==2.9.3
gunicorn==20.1.0
openai==0.11.0
django-cors-headers
dj_database_url
python-dotenv